Enum samd21_mini::atsamd21g18a::gclk::clkctrl::IDR

pub enum IDR {
    DFLL48,
    FDPLL,
    FDPLL32K,
    WDT,
    RTC,
    EIC,
    USB,
    EVSYS_0,
    EVSYS_1,
    EVSYS_2,
    EVSYS_3,
    EVSYS_4,
    EVSYS_5,
    EVSYS_6,
    EVSYS_7,
    EVSYS_8,
    EVSYS_9,
    EVSYS_10,
    EVSYS_11,
    SERCOMX_SLOW,
    SERCOM0_CORE,
    SERCOM1_CORE,
    SERCOM2_CORE,
    SERCOM3_CORE,
    SERCOM4_CORE,
    SERCOM5_CORE,
    TCC0_TCC1,
    TCC2_TC3,
    TC4_TC5,
    TC6_TC7,
    ADC,
    AC_DIG,
    AC_ANA,
    DAC,
    I2S_0,
    I2S_1,
    _Reserved(u8),
}

Possible values of the field ID

Variants
DFLL48

DFLL48

FDPLL

FDPLL

FDPLL32K

FDPLL32K

WDT

WDT

RTC

RTC

EIC

EIC

USB

USB

EVSYS_0

EVSYS_0

EVSYS_1

EVSYS_1

EVSYS_2

EVSYS_2

EVSYS_3

EVSYS_3

EVSYS_4

EVSYS_4

EVSYS_5

EVSYS_5

EVSYS_6

EVSYS_6

EVSYS_7

EVSYS_7

EVSYS_8

EVSYS_8

EVSYS_9

EVSYS_9

EVSYS_10

EVSYS_10

EVSYS_11

EVSYS_11

SERCOMX_SLOW

SERCOMX_SLOW

SERCOM0_CORE

SERCOM0_CORE

SERCOM1_CORE

SERCOM1_CORE

SERCOM2_CORE

SERCOM2_CORE

SERCOM3_CORE

SERCOM3_CORE

SERCOM4_CORE

SERCOM4_CORE

SERCOM5_CORE

SERCOM5_CORE

TCC0_TCC1

TCC0_TCC1

TCC2_TC3

TCC2_TC3

TC4_TC5

TC4_TC5

TC6_TC7

TC6_TC7

ADC

ADC

AC_DIG

AC_DIG

AC_ANA

AC_ANA

DAC

DAC

I2S_0

I2S_0

I2S_1

I2S_1

_Reserved(u8)

Reserved

Methods

impl IDR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_dfll48(&self) -> bool

Checks if the value of the field is DFLL48

pub fn is_fdpll(&self) -> bool

Checks if the value of the field is FDPLL

pub fn is_fdpll32k(&self) -> bool

Checks if the value of the field is FDPLL32K

pub fn is_wdt(&self) -> bool

Checks if the value of the field is WDT

pub fn is_rtc(&self) -> bool

Checks if the value of the field is RTC

pub fn is_eic(&self) -> bool

Checks if the value of the field is EIC

pub fn is_usb(&self) -> bool

Checks if the value of the field is USB

pub fn is_evsys_0(&self) -> bool

Checks if the value of the field is EVSYS_0

pub fn is_evsys_1(&self) -> bool

Checks if the value of the field is EVSYS_1

pub fn is_evsys_2(&self) -> bool

Checks if the value of the field is EVSYS_2

pub fn is_evsys_3(&self) -> bool

Checks if the value of the field is EVSYS_3

pub fn is_evsys_4(&self) -> bool

Checks if the value of the field is EVSYS_4

pub fn is_evsys_5(&self) -> bool

Checks if the value of the field is EVSYS_5

pub fn is_evsys_6(&self) -> bool

Checks if the value of the field is EVSYS_6

pub fn is_evsys_7(&self) -> bool

Checks if the value of the field is EVSYS_7

pub fn is_evsys_8(&self) -> bool

Checks if the value of the field is EVSYS_8

pub fn is_evsys_9(&self) -> bool

Checks if the value of the field is EVSYS_9

pub fn is_evsys_10(&self) -> bool

Checks if the value of the field is EVSYS_10

pub fn is_evsys_11(&self) -> bool

Checks if the value of the field is EVSYS_11

pub fn is_sercomx_slow(&self) -> bool

Checks if the value of the field is SERCOMX_SLOW

pub fn is_sercom0_core(&self) -> bool

Checks if the value of the field is SERCOM0_CORE

pub fn is_sercom1_core(&self) -> bool

Checks if the value of the field is SERCOM1_CORE

pub fn is_sercom2_core(&self) -> bool

Checks if the value of the field is SERCOM2_CORE

pub fn is_sercom3_core(&self) -> bool

Checks if the value of the field is SERCOM3_CORE

pub fn is_sercom4_core(&self) -> bool

Checks if the value of the field is SERCOM4_CORE

pub fn is_sercom5_core(&self) -> bool

Checks if the value of the field is SERCOM5_CORE

pub fn is_tcc0_tcc1(&self) -> bool

Checks if the value of the field is TCC0_TCC1

pub fn is_tcc2_tc3(&self) -> bool

Checks if the value of the field is TCC2_TC3

pub fn is_tc4_tc5(&self) -> bool

Checks if the value of the field is TC4_TC5

pub fn is_tc6_tc7(&self) -> bool

Checks if the value of the field is TC6_TC7

pub fn is_adc(&self) -> bool

Checks if the value of the field is ADC

pub fn is_ac_dig(&self) -> bool

Checks if the value of the field is AC_DIG

pub fn is_ac_ana(&self) -> bool

Checks if the value of the field is AC_ANA

pub fn is_dac(&self) -> bool

Checks if the value of the field is DAC

pub fn is_i2s_0(&self) -> bool

Checks if the value of the field is I2S_0

pub fn is_i2s_1(&self) -> bool

Checks if the value of the field is I2S_1