Struct samd21_mini::nvmctrl::ctrla::_CMDW

pub struct _CMDW<'a> { /* fields omitted */ }

Proxy

Methods

impl<'a> _CMDW<'a>

pub fn variant(self, variant: CMDW) -> &'a mut W

Writes variant to the field

pub fn er(self) -> &'a mut W

Erase Row - Erases the row addressed by the ADDR register.

pub fn wp(self) -> &'a mut W

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register.

pub fn ear(self) -> &'a mut W

Erase Auxiliary Row - Erases the auxiliary row addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

pub fn wap(self) -> &'a mut W

Write Auxiliary Page - Writes the contents of the page buffer to the page addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

pub fn sf(self) -> &'a mut W

Security Flow Command

pub fn wl(self) -> &'a mut W

Write lockbits

pub fn lr(self) -> &'a mut W

Lock Region - Locks the region containing the address location in the ADDR register.

pub fn ur(self) -> &'a mut W

Unlock Region - Unlocks the region containing the address location in the ADDR register.

pub fn sprm(self) -> &'a mut W

Sets the power reduction mode.

pub fn cprm(self) -> &'a mut W

Clears the power reduction mode.

pub fn pbc(self) -> &'a mut W

Page Buffer Clear - Clears the page buffer.

pub fn ssb(self) -> &'a mut W

Set Security Bit - Sets the security bit by writing 0x00 to the first byte in the lockbit row.

pub fn invall(self) -> &'a mut W

Invalidates all cache lines.

pub unsafe fn bits(self, value: u8) -> &'a mut W

Writes raw bits to the field