Struct samd21_mini::sercom::SPIMaster3

pub struct SPIMaster3 { /* fields omitted */ }

SPIMasterX represents the corresponding SERCOMX instance configured to act in the role of an SPI Master. Objects of this type implement the HAL FullDuplex and blocking SPI traits.

Methods

impl SPIMaster3

pub fn new<F>(
clock: &Sercom3CoreClock,
freq: F,
mode: Mode,
sercom: SERCOM3,
pm: &mut PM,
pinout: SPI3Pinout
) -> SPIMaster3 where
F: Into<Hertz>,

Power on and configure SERCOMX to work as an SPI Master operating with the specified frequency and SPI Mode. The pinout specifies which pins are bound to the MISO, MOSI, SCK functions.

pub fn free(self) -> (SPI3Pinout, SERCOM3)

Tear down the SPI instance and yield the constituent pins and SERCOM instance. No explicit de-initialization is performed.